Realizando ajustes no repository

+++ src/main/java/com/estudo/myfoodapi/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -5,9 +5,11 @@ import com.estudo.myfoodapi.domain.repository.CozinhaRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Objects;
 
 @Component
 public class CozinhaRepositoryImpl implements CozinhaRepository {
@@ -33,8 +35,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        cozinha = buscar(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+        if(Objects.isNull(cozinha)){
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinha);
     }
 }
